pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "garmin-data"
version = "0.1.0"
description = "Personal Garmin health and activity data pipeline with AI-ready insights"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
keywords = ["garmin", "health", "fitness", "data", "sqlite", "cli", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Health and Fitness",
]
dependencies = [
    "garminconnect>=0.3.3",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.scripts]
garmin-sync       = "integrations.garmin.cli:sync"
garmin-summary    = "integrations.garmin.cli:summary"
garmin-activities = "integrations.garmin.cli:activities"
garmin-insights   = "integrations.garmin.cli:insights"
garmin-ai-context = "integrations.garmin.cli:ai_context"

[tool.hatch.build.targets.wheel]
packages = ["integrations"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "freezegun>=1.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
